clean base file and refactor code
