+++ .gitignore
@@ -5,3 +5,4 @@ __pycache__/
 build/*
 dist/*
 .DS_Store
+.idea
